// apps/api/src/statistics/dto/statistics.dto.ts
export class DropoutRateDto {
  courseId: string;

  courseName: string;

  dropoutRate: number;

  dropoutCount: number;

  criticalPoints: {
    timePoint: number; // days since enrollment
    dropoutCount: number;
  }[];
}